+++ CHANGELOG.md
@@ -21,6 +21,7 @@ limitations under the License.
 - New: You can create client from existing client instance or model configuration to avoid loading model configuration from server
 - Fix: Leak of ModelClient during Triton creation
 - Fix: Fixed non-declared project dependencies (removed from use in code or added to package dependecies)
+- Fix: Remote model is being unloaded from Triton when RemoteTriton is closed.
 
 [//]: <> (put here on external component update with short summary what change or link to changelog)